feat(global)[NFE-12] Add more error handling for form fields

+++ src/main/java/com/mauricio/model/MainModel.java
@@ -4,13 +4,14 @@ import com.mauricio.domain.assinatura.CertificadoDigital;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.util.Enumeration;
 
 @Getter
 @Setter
 public class MainModel {
     private CertificadoDigital certificado;
-    private Integer nroLote;
+    private BigInteger nroLote;
     private String arquivoLoteSp;
     private Enumeration<String> aliases;
 }
